+++ app/Mailer/MailerServiceProvider.php
@@ -8,8 +8,8 @@ use CultuurNet\UDB3\Broadway\Domain\DomainMessageIsReplayed;
 use CultuurNet\UDB3\Container\AbstractServiceProvider;
 use CultuurNet\UDB3\Error\LoggerFactory;
 use CultuurNet\UDB3\Error\LoggerName;
-use CultuurNet\UDB3\Mailer\Ownership\SendMailsForOwnershipEventHandler;
-use CultuurNet\UDB3\Mailer\Ownership\SendOwnershipMailCommandHandler;
+use CultuurNet\UDB3\Mailer\Handler\SendMailsForOwnershipEventHandler;
+use CultuurNet\UDB3\Mailer\Handler\SendOwnershipMailCommandHandler;
 use CultuurNet\UDB3\Organizer\OrganizerServiceProvider;
 use CultuurNet\UDB3\Ownership\Repositories\Search\OwnershipSearchRepository;
 use CultuurNet\UDB3\User\UserIdentityResolver;
